Move inject_entropy_with_timestamp() inside mutex protected sections
inject_entropy_with_timestamp() is used to change the
entropy of the RNG each time we call tee_dispatch_open_session /
tee_dispatch_close_session. This is performed outside of a
session context, and outside mutex protection.
inject_entropy_with_timestamp() calls RPC to get the REE time.
When freing the shared memory in case of concurrent TA, we may
have a race condition between optee_os and the supplicant
when accessing / freeing the shared memory.

This patch calls inject_entropy_with_timestamp() inside the
mutex protected section, when calling tee_ta_init_session().
