+++ core/kernel/tee_dispatch.c
@@ -24,14 +24,11 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-
 #include <kernel/tee_dispatch.h>
-
 #include <kernel/tee_ta_manager.h>
 #include <mm/core_memprot.h>
 #include <mm/core_mmu.h>
 #include <kernel/tee_compat.h>
-#include <kernel/tee_time.h>
 #include <tee/tee_cryp_utl.h>
 
 /* Sessions opened from normal world */
@@ -81,14 +78,6 @@ static TEE_Result update_clnt_id(const TEE_Identity *in, TEE_Identity *out)
 	return TEE_SUCCESS;
 }
 
-static void inject_entropy_with_timestamp(void)
-{
-	TEE_Time current;
-
-	if (tee_time_get_sys_time(&current) == TEE_SUCCESS)
-		tee_prng_add_entropy((uint8_t *)&current, sizeof(current));
-}
-
 TEE_Result tee_dispatch_open_session(struct tee_dispatch_open_session_in *in,
 				     struct tee_dispatch_open_session_out *out)
 {
@@ -116,13 +105,6 @@ TEE_Result tee_dispatch_open_session(struct tee_dispatch_open_session_in *in,
 	memcpy(out->params, in->params, sizeof(in->params));
 	update_out_param(&param, out->params);
 
-	/*
-	 * The occurrence of open/close session command is usually
-	 * un-predictable, using this property to increase randomness
-	 * of prng
-	 */
-	inject_entropy_with_timestamp();
-
 cleanup_return:
 	if (res != TEE_SUCCESS)
 		DMSG("  => Error: %x of %d", (unsigned int)res, (int)res_orig);
@@ -134,8 +116,6 @@ cleanup_return:
 
 TEE_Result tee_dispatch_close_session(struct tee_close_session_in *in)
 {
-	inject_entropy_with_timestamp();
-
 	return tee_ta_close_session((struct tee_ta_session *)in->sess,
 				    &tee_open_sessions, NSAPP_IDENTITY);
 }
